requirements.txt
# Kiro-2API Dependencies
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
httpx>=0.25.0,<1.0.0
loguru>=0.7.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic-settings>=2.0.0,<3.0.0
python-multipart>=0.0.6,<1.0.0
tiktoken>=0.5.0,<1.0.0
